requirements.txt
PySide6
pillow
numpy
torch
